+++ src/components/Card/card.scss
@@ -1,4 +1,4 @@
-@use './utilities/variables';
+@use '../../styles/_variables';
 
 section {
   background-color: variables.$white;
