+++ src/styles/_global.scss
@@ -1,3 +1,7 @@
 * {
   background-color: hsl(275, 100%, 97%);
 }
+
+html {
+  height: 100%
+}
