--- src/styles/app.scss
@@ -1,3 +1,0 @@
-.card-and-header-section {
-  border: 1px solid red;
-}
